+++ src/utils/enums/marketplace-shot-name.enum.ts
@@ -1,0 +1,5 @@
+export enum MarketplaceShotName {
+  SIDEBAR = 'sidebar',
+  PREVIEW = 'preview',
+  SIGN_IN = 'signin',
+}